+++ mvn/mvn.sh
@@ -60,7 +60,7 @@ function mrcmd_plugins_mvn_method_install() {
 }
 
 function mrcmd_plugins_mvn_method_uninstall() {
-  if [ -f "${MVN_SETTINGS_PATH}" ] && [[ "${MVN_CONFIG_DIR}" == "${APPX_DIR}/.m2" ]] ; then
+  if [ -f "${MVN_SETTINGS_PATH}" ] && [[ "${MVN_CONFIG_DIR}" == "${APPX_DIR}/.m2" ]]; then
     mrcore_lib_rmdir "${MVN_CONFIG_DIR}"
   fi